refactor: 단골 도메인 CursorResult DTO 공용 페이지네이션 적용
